src/lambda2/requirements.txt
boto3
python-dotenv
aws-lambda-typing
